requirements.txt
# GEEWONI Trading System Requirements
# Python 3.12

# Core
streamlit==1.38.0
yfinance==0.2.40
python-telegram-bot==21.4
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
openai==1.51.2
pytz==2024.1
tiktoken==0.5.1
nest_asyncio==1.6.0

# Phase 1: Real-time data (optional - bot works with Yahoo only if not set)
finnhub-python==2.4.19

# Phase 3: Dashboard charts
plotly==5.18.0

# 🆕 Scheduling & Automation
apscheduler==3.10.4

# 🆕 News & RSS
feedparser==6.0.10
beautifulsoup4==4.12.3

# Optional: News API
# newsapi-python==0.2.7

# IBKR Integration (coming soon)
# ib_insync==0.9.86